refactor: refactor ddl manager
