+++ src/frontend/src/instance/distributed.rs
@@ -32,7 +32,7 @@ use chrono::DateTime;
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
 use common_catalog::format_full_table_name;
 use common_error::ext::BoxedError;
-use common_meta::ddl_manager::{Context, DdlExecutorRef};
+use common_meta::ddl::{DdlExecutorRef, ExecutorContext};
 use common_meta::key::schema_name::{SchemaNameKey, SchemaNameValue};
 use common_meta::rpc::ddl::{DdlTask, SubmitDdlTaskRequest, SubmitDdlTaskResponse};
 use common_meta::rpc::router::{Partition, Partition as MetaPartition};
@@ -460,7 +460,7 @@ impl DistInstance {
         };
 
         self.ddl_executor
-            .submit_ddl_task(&Context, req)
+            .submit_ddl_task(&ExecutorContext::default(), req)
             .await
             .context(error::ExecuteDdlSnafu)?;
 
@@ -485,7 +485,7 @@ impl DistInstance {
         };
 
         self.ddl_executor
-            .submit_ddl_task(&Context, request)
+            .submit_ddl_task(&ExecutorContext::default(), request)
             .await
             .context(error::ExecuteDdlSnafu)
     }
@@ -505,7 +505,7 @@ impl DistInstance {
         };
 
         self.ddl_executor
-            .submit_ddl_task(&Context, request)
+            .submit_ddl_task(&ExecutorContext::default(), request)
             .await
             .context(error::ExecuteDdlSnafu)
     }
@@ -519,7 +519,7 @@ impl DistInstance {
         };
 
         self.ddl_executor
-            .submit_ddl_task(&Context, request)
+            .submit_ddl_task(&ExecutorContext::default(), request)
             .await
             .context(error::ExecuteDdlSnafu)
     }
