+++ src/meta-srv/src/metasrv.rs
@@ -22,6 +22,7 @@ use api::v1::meta::Peer;
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
 use common_greptimedb_telemetry::GreptimeDBTelemetryTask;
 use common_grpc::channel_manager;
+use common_meta::ddl::DdlExecutorRef;
 use common_meta::key::TableMetadataManagerRef;
 use common_procedure::options::ProcedureConfig;
 use common_procedure::ProcedureManagerRef;
@@ -33,7 +34,6 @@ use snafu::ResultExt;
 use tokio::sync::broadcast::error::RecvError;
 
 use crate::cluster::MetaPeerClientRef;
-use crate::ddl::DdlManagerRef;
 use crate::election::{Election, LeaderChangeMessage};
 use crate::error::{RecoverProcedureSnafu, Result};
 use crate::handler::HeartbeatHandlerGroup;
@@ -97,7 +97,7 @@ impl MetaSrvOptions {
     }
 }
 
-pub struct MetasrvInfo {
+pub struct MetaSrvInfo {
     pub server_addr: String,
 }
 
@@ -191,7 +191,7 @@ pub struct MetaSrv {
     procedure_manager: ProcedureManagerRef,
     metadata_service: MetadataServiceRef,
     mailbox: MailboxRef,
-    ddl_manager: DdlManagerRef,
+    ddl_executor: DdlExecutorRef,
     table_metadata_manager: TableMetadataManagerRef,
     greptimedb_telemetry_task: Arc<GreptimeDBTelemetryTask>,
     pubsub: Option<(PublishRef, SubscribeManagerRef)>,
@@ -342,8 +342,8 @@ impl MetaSrv {
         &self.mailbox
     }
 
-    pub fn ddl_manager(&self) -> &DdlManagerRef {
-        &self.ddl_manager
+    pub fn ddl_executor(&self) -> &DdlExecutorRef {
+        &self.ddl_executor
     }
 
     pub fn procedure_manager(&self) -> &ProcedureManagerRef {
